rename chart component, write additional filter logic
