+++ pages/index.js
@@ -10,14 +10,6 @@ import {
   StyledFlexColumnWrapper,
 } from "@/SharedStyledComponents";
 import Tooltip from "@/components/Tooltip";
-import Loader from "@/components/Loader";
-import dynamic from "next/dynamic";
-import ErrorMessage from "@/components/ErrorMessage";
-
-const TensionChart = dynamic(() => import("../components/TensionChart"), {
-  ssr: false,
-  loading: () => <Loader itemText="... loading" />,
-});
 
 const StyledTensionForm = styled(StyledForm)`
   margin: 1rem;
@@ -45,12 +37,6 @@ const StyledMessage = styled.p`
   text-align: center;
   font-weight: 600;
   margin: 1rem auto;
-`;
-
-const StyledGraphButton = styled(StyledButton)`
-  width: fit-content;
-  padding: 0.5rem;
-  border-style: none;
 `;
 
 const StyledButtonWrapper = styled(StyledWrapper)`
@@ -83,16 +69,10 @@ const StyledAddDetailsLink = styled(StyledStandardLink)`
     $actionButton ? "1px solid black" : "none"};
 `;
 
-export default function HomePage({
-  onAddEmotionEntry,
-  handleToggleTooltip,
-  emotionEntries,
-  theme,
-}) {
+export default function HomePage({ onAddEmotionEntry, handleToggleTooltip }) {
   const [isFormSubmitted, setIsFormSubmitted] = useState(false);
   const [id, setId] = useState();
   const [tension, setTension] = useState(0);
-  const [chartIsShown, setChartIsShown] = useState(true);
 
   function handleSubmit(event) {
     event.preventDefault();
@@ -104,26 +84,6 @@ export default function HomePage({
     onAddEmotionEntry(data, newId);
     setId(newId);
     setIsFormSubmitted(!isFormSubmitted);
-  }
-
-  //logic for Graph
-  const currentShortDate = new Date().toISOString().slice(0, 10);
-  function compare(a, b) {
-    if (a.isoDate < b.isoDate) {
-      return -1;
-    }
-    if (a.isoDate > b.isoDate) {
-      return 1;
-    }
-    return 0;
-  }
-  const filteredData = emotionEntries
-    .filter((entry) => currentShortDate === entry.isoDate?.slice(0, 10))
-    .sort(compare);
-  const xValues = filteredData.map((entry) => entry.timeAndDate.slice(-5));
-  const yValues = filteredData.map((entry) => entry.tensionLevel);
-  function handleChart() {
-    setChartIsShown(!chartIsShown);
   }
 
   return (
@@ -181,22 +141,6 @@ export default function HomePage({
             </>
           )}
         </StyledTensionForm>
-
-        {chartIsShown && xValues.length != 0 ? (
-          <TensionChart
-            emotionEntries={emotionEntries}
-            theme={theme}
-            xValues={xValues}
-            yValues={yValues}
-            title="Daily Tension Graph"
-          />
-        ) : (
-          <ErrorMessage itemText="No Data for Graph" />
-        )}
-
-        <StyledGraphButton type="button" onClick={handleChart}>
-          {chartIsShown === true ? "Hide chart" : "Show chart"}
-        </StyledGraphButton>
       </StyledFlexColumnWrapper>
     </>
   );
